auth is working!
This commit adds session based auth for all api endpoints.
If the user is not signed in, the front end won't be loaded either.

Auth is configured to serve read only endpoints to anyone, even if they
aren't authenticated. This is because the front end GET calls need to be
adapted so they pass in the csrf_token like the POST fetch calls do,
otherwise the GET calls would break.

POST endpoints require the user to be
authenticated though. They are configured and working this way.

Unit tests broke because of auth, but this commit also fixes them