Adjust mark as exited endpoint to allow null permanent_housing value
It previously only allowed true or false.
The unit tests and the docs page have both been updated.

Regarding issue #60, I couldn't reproduce the issue. I tried changing
the value of permanent_housing for a visit, and I watched the JSON
change from "true" to "false" to "null" as expected.